% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{catch}
\alias{catch}
\title{Turn on error catching}
\usage{
catch()
}
\value{
used for side effects
}
\description{
Sometimes iterations diverge, leading to errors. If error catching is turned on, the iteration will return a warning and the last non-error value.
}
\examples{
f <- function(x) if (x > 0) { x-2*x} else { stop('aargh') }
\dontrun{
g <- fp(f)
g(3)
}
g <- fp(f)+catch()
g(3)
}
\author{
Simon Barthelme
}

