% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{convergence}
\alias{convergence}
\title{Change convergence criterion in fixed point iteration}
\usage{
convergence(fun, tol, maxIter)
}
\arguments{
\item{fun}{convergence check: either a function with two arguments or a character string like "rmse"}

\item{tol}{tolerance}

\item{maxIter}{}
}
\value{
used for side effects
}
\description{
A fixed point is defined by f(x)=x. Often this condition can't be fulfilled exactly, and so we must check convergence using a looser criterion, like abs(f(x)-x) < tol.
The convergence criterion can be specified using a function of two arguments: d(x,x'), returning either a logical value or a positive scalar, to be checked against the tolerance 'tol'. 
The predefined convergence functions are:
mse=function(a,b) mean((a-b)^2)
rmse=function(a,b) sqrt(mean((a-b)^2))
mad=function(a,b) mean(abs(a-b))
mre=function(a,b) mean(abs(a-b)/abs(a+b)
If convergence can't be reached at all, we interrupt the search after a maximum number of iterations (set by maxIter).
}
\examples{
f <- function(x) x/2
g <- fp(f)
g(3) #by default convergence is assessed using all.equal
(g+convergence("rmse",tol=.1))(3) #root-mean-square less than .1
(g+convergence("rmse",tol=.01))(3) #root-mean-square less than .01
#Custom function
d = function(a,b) sum(abs(log(a)-log(b)))
(g+convergence(d,tol=.01))(3) #can't converge because fixed point is at 0
}
\author{
Simon Barthelme
}

