% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{fp}
\alias{fp}
\title{Turn a function into a fixed point iteration}
\usage{
fp(f)
}
\arguments{
\item{f}{a function}
}
\value{
g another function
}
\description{
Given a function f, return g such that g(x) is a fixed point of f (or at least an approximate one). 
Generally speaking the idea will be iterate f(f(f(f(....f(x))))) until a fixed point is reached, which of course may fail entirely if there are no stable fixed points.
}
\examples{
#This function has a fixed point at x = 0, i.e. 0/2 = 0
f <- function(x) x/2
g <- fp(f)
g
g(3) #The fixed point is only approximate.
g(-2) 
#The default criterion is all.equal(f(x),x). Check:
all.equal(g(3),f(g(3)))
}
\author{
Simon Barthelme
}
\seealso{
iter (for fixed length iterations), viewer (for viewing intermediate results), store (for storing them)
}

