% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{iter}
\alias{iter}
\title{Turn a function into an iterated mapping}
\usage{
iter(f)
}
\arguments{
\item{f}{a function}
}
\value{
g another function
}
\description{
Given a function f, return g such that g(x,k) = { f^k } (x) = f(f(f(f(...f(x)))))
}
\examples{
f <- function(x) x*2
g <- iter(f)
g
g(3,nIter=2)  ## Corresponds to:
f(3) \%>\% f
g(3,nIter=3) ## Corresponds to: 
f(3) \%>\% f \%>\% f
}
\author{
Simon Barthelme
}
\seealso{
fp (for fixed points), viewer (for viewing intermediate results), store (for storing them)
}

