% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{store}
\alias{store}
\alias{stored}
\title{Store intermediate iterations}
\usage{
store(fun, every = 1)

stored(x)
}
\arguments{
\item{fun}{a function}

\item{every}{frequency: run every k iteration (default 1, meaning at every iteration)}
}
\value{
used for side effects
}
\description{
Store intermediate values in a function iteration. Argument "fun" must be a function with at most two arguments. It will be passed the current value and the iteration number, and it must return an object that will be stored. See examples
}
\section{Functions}{
\itemize{
\item \code{stored}: extracted stored values
}}
\examples{
f <- function(x) x/2
g <- iter(f)+store(function(x) x)
res <- g(3,5)
stored(res) #The result of 5 iterations, as a vector
#Store more info:
st <- function(x,ind) list(value=x,iter=ind)
res <- (iter(f)+store(st))(3,5)
stored(res) ## We now have a list of lists
#To extract elements, use map_ functions from package purrr
stored(res) \%>\% purrr::map_int("iter")
stored(res) \%>\% purrr::map_dbl("value")
stored(res) \%>\% purrr::map_df(identity)
#Store a summary
st <- function(x,ind) list(mvalue=mean(x),iter=ind)
g <- iter(f)+store(st)
g(rnorm(10),5) \%>\% stored \%>\% purrr::map_df(identity)
}
\author{
Simon Barthelme
}
\seealso{
viewer,stored
}

