% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{viewer}
\alias{viewer}
\title{Output intermediate results in iterations}
\usage{
viewer(fun, every = 1, print.iter = FALSE)
}
\arguments{
\item{fun}{a viewer function}

\item{every}{frequency: run every k iteration (default 1, meaning at every iteration)}

\item{print.iter}{print iteration number}
}
\value{
used for side effects
}
\description{
Add a viewer to an iteration, for monitoring progress. Argument "fun" must be a function with at most two arguments. It will be passed the current value and the iteration number.
Some default viewing functions are already defined like "print".
}
\examples{
f <- function(x) x/2
g <- fp(f)
(g+viewer(print))(.1)
# Now print only after 10 iterations
(g+viewer(print,every=10))(3)
# Print iteration number
(g+viewer(print,every=5,TRUE))(3)

# Define custom viewer:
v = function(x) print(x,digits=2)
(g+viewer(v,every=5,TRUE))(3)
v2 = function(x,ind) sprintf("At iteration \%i. Current value \%.3f",ind,x) \%>\% print
(g+viewer(v2,every=5))(3)
#You can have multiple viewers, running at different frequencies
v1 = function(x,ind) sprintf("At iteration \%i",ind)\%>\%print
(g+viewer(v1)+viewer(v2,every=5))(3)
#You can also use plotting
x0 <- rnorm(40)
plot(x0)
v <- function(x) points(x,col="grey")
(g+viewer(v))(x0)
}
\author{
Simon Barthelme
}

